requirements.txt
click==8.1.3
colorama==0.4.6
Pillow==9.4.0
pip==22.3.1
progressbar==2.5
PySimpleGUI==4.60.4
setuptools==67.0.0
